Add GetRandomQuoteQueryHandler
