+++ src/WiseReminder.Application/Quotes/GetRandomQuote/GetRandomQuoteQuery.cs
@@ -1,22 +1,5 @@
-﻿using AutoMapper;
-using WiseReminder.Application.Abstractions.MediatR;
-using WiseReminder.Domain.Abstractions;
-using WiseReminder.Domain.Quotes;
+﻿using WiseReminder.Application.Abstractions.MediatR;
 
 namespace WiseReminder.Application.Quotes.GetRandomQuote;
 
 public sealed record GetRandomQuoteQuery : IQuery<QuoteVm>;
-
-public sealed class GetRandomQuoteQueryHandler(IQuoteRepository quoteRepository, IMapper mapper)
-    : IQueryHandler<GetRandomQuoteQuery, QuoteVm>
-{
-    private readonly IMapper _mapper = mapper;
-    private readonly IQuoteRepository _quoteRepository = quoteRepository;
-
-    public async Task<Result<QuoteVm>> Handle(GetRandomQuoteQuery request, CancellationToken cancellationToken)
-    {
-        var quote = await _quoteRepository.GetRandomQuote();
-
-        return Result<QuoteVm>.Success(_mapper.Map<QuoteVm>(quote));
-    }
-}
